[FIX] Passing groups_columns in final fit of IAML.fit() function.
When passing groups DataFrame, X is already pruned of groups columns.
So the last fits (line 289) We are passing already pruned X.
When passing groups columns, X is pruned inside the Dataset.__init__()
So the last fits (line 289) stil have the groups columns.

Maybe we could refacto IAMLPipeline.fit and IAMLPipeline.fit_transform, to take a Dataset instead of raw X and Y ?